refactor(cli): inline cli-core into @lobu/cli
Merges packages/cli-core back into packages/cli/src/internal/. cli-core
had exactly one consumer (@lobu/cli) and no external users — it was
split out in #282 with the intent of eventually sharing helpers with
owletto-cli, but that integration never happened and the env-loader
contracts don't line up cleanly anyway.

Removing one published package from the release:
- One less npm OIDC trusted-publisher registration to maintain.
- @lobu/cli ships self-contained; end users install one package instead
  of one+transitive.
- Same source-level organization preserved via packages/cli/src/internal/.

Drops cli-core from publish-packages.mjs, bump-version.mjs,
release-please-config.json, tsconfig paths, root build scripts, and
every Dockerfile stub. CLI builds and --help works end-to-end.
