+++ docker/app/Dockerfile
@@ -36,8 +36,8 @@ COPY packages/owletto-web/package.jso[n] packages/owletto-web/
 
 # Stub out workspace packages we don't build in this image so bun install doesn't
 # error on missing manifests.
-RUN mkdir -p packages/cli packages/cli-core packages/landing packages/owletto-cli packages/owletto-extension packages/mcpsandbox \
-    && for p in cli cli-core landing owletto-extension mcpsandbox; do \
+RUN mkdir -p packages/cli packages/landing packages/owletto-cli packages/owletto-extension packages/mcpsandbox \
+    && for p in cli landing owletto-extension mcpsandbox; do \
         echo "{\"name\":\"@lobu/$p\",\"version\":\"0.0.0\",\"private\":true}" > "packages/$p/package.json"; \
       done \
     && echo '{"name":"owletto","version":"0.0.0","private":true}' > packages/owletto-cli/package.json
