+++ docker/embeddings-service/Dockerfile
@@ -19,11 +19,11 @@ COPY patches/ ./patches/
 COPY packages/owletto-embeddings/package.json packages/owletto-embeddings/
 
 # Stub workspaces not needed at build time
-RUN mkdir -p packages/cli packages/cli-core packages/core packages/gateway packages/landing \
+RUN mkdir -p packages/cli packages/core packages/gateway packages/landing \
                packages/owletto-backend packages/owletto-cli packages/owletto-connectors \
                packages/owletto-extension packages/owletto-openclaw packages/owletto-sdk \
                packages/owletto-web packages/owletto-worker packages/worker packages/mcpsandbox \
-    && for p in cli cli-core core gateway landing owletto-backend owletto-connectors \
+    && for p in cli core gateway landing owletto-backend owletto-connectors \
                 owletto-extension owletto-openclaw owletto-sdk owletto-web owletto-worker worker mcpsandbox; do \
         echo "{\"name\":\"@lobu/$p\",\"version\":\"0.0.0\",\"private\":true}" > "packages/$p/package.json"; \
       done \
