+++ package.json
@@ -7,7 +7,7 @@
     "packages/*"
   ],
   "scripts": {
-    "build": "cd packages/core && bun run build && cd ../cli-core && bun run build && cd ../gateway && bun run build && cd ../worker && bun run build && cd ../..",
+    "build": "cd packages/core && bun run build && cd ../gateway && bun run build && cd ../worker && bun run build && cd ../..",
     "format": "biome format --config-path config/biome.config.json --write .",
     "format:check": "biome format --config-path config/biome.config.json .",
     "lint": "biome lint --config-path config/biome.config.json .",
@@ -21,12 +21,12 @@
     "jscpd": "bunx jscpd --pattern \"packages/**/*.ts\" --reporters \"console\" --ignore \"**/node_modules/**,**/dist/**,**/coverage/**,**/.bun/**,**/.turbo/**,workspaces/**,charts/**,integration-tests/**,bin/**,**/__tests__/**,**/__mocks__/**,**/*.d.ts\"",
     "depcruise": "bunx --bun depcruise --config config/dependency-cruiser.config.cjs packages",
     "dev": "docker compose -f docker/docker-compose.yml watch",
-    "build:packages": "cd packages/core && bun run build && cd ../cli-core && bun run build && cd ../gateway && bun run build && cd ../worker && bun run build && cd ../cli && bun run build",
+    "build:packages": "cd packages/core && bun run build && cd ../gateway && bun run build && cd ../worker && bun run build && cd ../cli && bun run build",
     "build:owletto": "cd packages/owletto-sdk && bun run build && cd ../owletto-embeddings && bun run build && cd ../owletto-cli && bun run build && cd ../..",
-    "watch:packages": "tsc -b --watch packages/core packages/cli-core packages/gateway packages/worker",
+    "watch:packages": "tsc -b --watch packages/core packages/gateway packages/worker",
     "test:packages": "cd packages/core && bun run test && cd ../gateway && bun run test && cd ../worker && bun run test",
     "test:owletto": "cd packages/owletto-cli && bun run runtime:prepare && cd ../..",
-    "typecheck:packages": "cd packages/core && bun run typecheck && cd ../cli-core && bun run typecheck && cd ../gateway && bun run typecheck && cd ../worker && bun run typecheck",
+    "typecheck:packages": "cd packages/core && bun run typecheck && cd ../gateway && bun run typecheck && cd ../worker && bun run typecheck",
     "typecheck:owletto": "cd packages/owletto-backend && bun run typecheck && cd ../owletto-cli && bun run typecheck && cd ../owletto-embeddings && bun run typecheck && cd ../owletto-sdk && bun run typecheck && cd ../..",
     "knip": "bunx knip --config config/knip.ts",
     "gitleaks": "docker run --rm -v $(pwd):/repo -w /repo zricethezav/gitleaks:latest detect --source=/repo --no-git --redact --config=/repo/config/gitleaks.toml",
