--- packages/cli-core/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "@lobu/cli-core",
-  "version": "4.0.0",
-  "license": "Apache-2.0",
-  "type": "module",
-  "description": "Shared CLI internals for @lobu/cli — local-config loader, credential store, context resolution",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "bun": "./src/index.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.js"
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsc",
-    "typecheck": "tsc --noEmit",
-    "clean": "rm -rf dist"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/lobu-ai/lobu.git",
-    "directory": "packages/cli-core"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "engines": {
-    "node": ">=18.0.0"
-  }
-}
