+++ packages/cli/package.json
@@ -26,7 +26,6 @@
     "directory": "packages/cli"
   },
   "dependencies": {
-    "@lobu/cli-core": "workspace:*",
     "@lobu/core": "workspace:*",
     "chalk": "^5.3.0",
     "commander": "^12.0.0",
